Replace Chrome-specific extension code with WebExtension-compliant code
By making the messaging and other browser-level operations compliant to
the WebExtension spec, we *should* be able to target Firefox and Edge
without too much drama.

In order to retain compatibility with Chrome, the
'webextension-polyfill' library is included in the build. This is
injected via `manifest.json`, as well as the panel page using a
`<script>` tag. The polyfill is designed to be a no-op under browsers
that already implement to the WebExtension spec, so no conditional build
flags should be required.