Prevent/trap dependency trace failures
Fixes #23

When selecting an 'Init' message, some uncaught errors would cause the
Panel to appear unresponsive.

First, coerce `msg.data` to an object even when missing (as is the case
with Init messages and possibly other types) to prevent `deepPick` from
throwing.

Further, if the underlying eval throws an exception, then log it to the
console and return an empty trace result.

The eval fails and Unit Tests do not display the correct content
because the Message is described as the string `'Init (ContainerName)'`
which is not accurate as far as Casium is concerned. I'd like to refine
how Init messages are reported to the DevTools, but I'll work on that
once the currently overhaul as part of the Electron application is
completed.