Fix issues with deep model diffs and array changes
When viewing a model diff where a modified property lived on the
'boundary' of the default expansion depth of '3', rendering errors would
occur. Increasing the default expansion depth to 100 for model diffs
acts as a (hacky) workaround.

Also fixes rendering of array changes; the format for unchanged array
values was incorrectly assumed to be `[['+', 'newValue'], [' ',
'unchangedValue']]`, when in fact it should be `[['+', 'newValue'], ['
']]`.`