Use upcoming 'instrumentation' API
Instead of relying on a `_ARCH_DEV_TOOLS_STATE` global in the inspected
page, inject a script into the context of the inspected page that uses
the instrumentation API to retrieve the State Manager and subscribe to
message dispatch.

The Instrumenter class offers a pluggable 'backend' mechanism to make
it easier to implement different types of client dev tool applications
in the future.

For now, the 'WebExtension' backend operates in a very similar manner to
the previous implementation; it uses a combination of
`browser.runtime.connect` and `window.postMessage` to provide messaging
between the inspected page and the DevTools panel.