Delta-based message storage for reduced memory usage
When working with large, complex applications, storing the previous and
next state of an application across many messages quickly consumes very
large amounts of memory.

In order to combat this, instead of storing the previous and next state
of an application, we instead use `json-delta` to create only store the
*changes* induced by a Message.

In addition to changing the storage to be delta-based, this commit also
removes the rather ugly customisation of `react-inspector`, replacing
it with a custom renderer function for the 'copy to clipboard'
behaviour, and delegating the rendering of the effective delta to the
`react-json-delta-inspector` component.