Only allow content scripts to run on 'top' frame.
When attempting to inspect a page that contains an iframe, the content script will be executed for each iframe on the page. This has the undesierable effect of 'breaking' the connection between the DevTools panel and the page being inspected. In the unlikely event that people want to inspect an application that has multiple instances of Casium running in both the 'top' frame and child frames, we can revisit this strategy.