Fix message log import/export
After the recent refactor to use a diff-based mechanism for state
changes, the message log import/export was sadly neglected. Now, the
initial state as well as a list of messages are exported.

In an effort to make the message log format more useful in the future,
I've also added a 'version' header, which is currently version '1'.
This gives us room to move around if we decide to change the message
log structure again.